// superlista-front/src/openapi/model/diaComidaDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RecetaDto } from './recetaDto';


export interface DiaComidaDto { 
    id?: number;
    dia?: DiaComidaDto.DiaEnum;
    categoria?: DiaComidaDto.CategoriaEnum;
    recetas?: Array<RecetaDto>;
}
export namespace DiaComidaDto {
    export type DiaEnum = 'LUNES' | 'MARTES' | 'MIERCOLES' | 'JUEVES' | 'VIERNES' | 'SABADO' | 'DOMINGO';
    export const DiaEnum = {
        Lunes: 'LUNES' as DiaEnum,
        Martes: 'MARTES' as DiaEnum,
        Miercoles: 'MIERCOLES' as DiaEnum,
        Jueves: 'JUEVES' as DiaEnum,
        Viernes: 'VIERNES' as DiaEnum,
        Sabado: 'SABADO' as DiaEnum,
        Domingo: 'DOMINGO' as DiaEnum
    };
    export type CategoriaEnum = 'DESAYUNO' | 'ALMUERZO' | 'COMIDA' | 'MERIENDA' | 'CENA';
    export const CategoriaEnum = {
        Desayuno: 'DESAYUNO' as CategoriaEnum,
        Almuerzo: 'ALMUERZO' as CategoriaEnum,
        Comida: 'COMIDA' as CategoriaEnum,
        Merienda: 'MERIENDA' as CategoriaEnum,
        Cena: 'CENA' as CategoriaEnum
    };
}