// superlista-front/src/openapi/model/menuSemanalCreateDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface MenuSemanalCreateDto { 
    categorias?: Array<MenuSemanalCreateDto.CategoriasEnum>;
}
export namespace MenuSemanalCreateDto {
    export type CategoriasEnum = 'DESAYUNO' | 'ALMUERZO' | 'COMIDA' | 'MERIENDA' | 'CENA';
    export const CategoriasEnum = {
        Desayuno: 'DESAYUNO' as CategoriasEnum,
        Almuerzo: 'ALMUERZO' as CategoriasEnum,
        Comida: 'COMIDA' as CategoriasEnum,
        Merienda: 'MERIENDA' as CategoriasEnum,
        Cena: 'CENA' as CategoriasEnum
    };
}